# kubilitics-backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Kubilitics Backend API
  description: |
    Production-grade Kubernetes topology and management API.
    
    This API provides comprehensive Kubernetes cluster management, real-time topology visualization,
    resource operations, and WebSocket streaming for live updates.
    
    **Key Features:**
    - Multi-cluster management
    - Real-time topology graph generation with relationship inference
    - WebSocket streaming for live updates
    - Resource CRUD operations
    - Logs, metrics, and events streaming
    - Export topology in multiple formats (PNG, PDF, SVG, JSON)
    
  version: 1.0.0
  contact:
    name: Kubilitics Team
    url: https://github.com/kubilitics
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: http://localhost:8080/api/v1
    description: Production server (configured per deployment)
...
paths:
...
  /ws/resources:
    get:
      summary: WebSocket connection for resource updates
      description: |
        Establish WebSocket connection to receive real-time Kubernetes resource updates.
        
        **Connection URL**: `ws://localhost:8080/ws/resources?cluster_id={id}&namespace={ns}`
        
        **Message Format**:
        ```json
        {
          "type": "RESOURCE_ADDED" | "RESOURCE_UPDATED" | "RESOURCE_DELETED",
          "cluster_id": "cluster-uuid",
          "resource": {
            "kind": "Pod",
            "namespace": "default",
            "name": "nginx-abc123",
            "data": { ... }
          },
          "timestamp": "2024-01-01T10:00:00Z"
        }
        ```
      tags:
        - WebSocket
      parameters:
        - name: cluster_id
          in: query
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Filter updates by namespace
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '400':
          description: Bad Request - Invalid parameters
        '401':
          description: Unauthorized
...